Escape values for commands that define a message
A not-so-well-documented aspect of commands like set-env and set-output
is that their value must be escaped in a specific way.

In ec29893 this change was made for set-output, however the escaping is
required for all commands which define a message:

https://github.com/actions/toolkit/blob/9ad01e4fd30025e8858650d38e95cfe9193a3222/packages/core/src/command.ts#L74

Therefore we define escapeData to match the core JavaScript toolkit
equivalent, and call that as required for commands which define
messages.

We add a small table test for set-env in the process.

A TODO for a later change would be to mirror more closely the toolkit
implementation of Command with a struct type that encodes (via its
methods) this requirement on encoding messages and any other nuances we
might have missed.