Define a generic API to issue arbitrary action commands
Goal here is to closely shadow the TypeScript counterpart.

https://github.com/actions/toolkit/blob/main/packages/core/src/command.ts

Essentially, a command is divided in 3 parts: name, message and
properties. All three parts are optional in the TypeScript API.
Properties are comma separated 'key=val' pairs accepted by some
commands.

e.g. ::name properties::message

Moreover, data type of 'message' and 'val' (in properties) is not
limited to 'string'. The TypeScript counterpart accepts generic types
for these arguments. A common function 'toCommandValue' is defined in
the TS API which JSON encodes non-string values for 'message' and 'val'.
Another function 'escapeProperty' is used to sanitize 'val' part of the
properties.

This commit proposes all these changes and migrates the existing APIs
to use them.