Panic on errors and remove deprecated functions (#41)
Previously, if writing to a env/path file failed, the library would silently fail if writing to the file failed. Similarly, if writing to an output stream fails, the library would silently fail. This updates the behavior to panic on these low-level errors.

Additionally, this removes deprecated ways of setting environment variables and paths given the new GitHub runners have been available for months. It also removes deprecated library functions.